backend/requirements.txt
# ── Web framework ──────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Supabase client ─────────────────────────────────
supabase==2.4.6

# ── Auth / JWT ─────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── File handling ───────────────────────────────────
python-multipart==0.0.9
pillow==10.3.0
pdf2image==1.17.0
pypdf==4.2.0

# ── OCR ────────────────────────────────────────────
easyocr==1.7.1
numpy<2

# ── AI / NLP (lightweight; swap for OpenAI if preferred) ──
transformers==4.40.2
torch==2.3.0

# ── Data validation ─────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0

# ── Environment ─────────────────────────────────────
python-dotenv==1.0.1

# ── HTTP client (used internally by supabase-py) ───
httpx==0.27.0

# ── Dev tooling (optional) ──────────────────────────
pytest==8.2.0
httpx==0.27.0          # also serves as test client
